+++ product/warehouse/src/webapp/config/webpack.prod.js
@@ -32,7 +32,9 @@ module.exports = webpackMerge(commonConfig, {
                 keep_fnames: true
             }
         }),
-        new ExtractTextPlugin('[name].[hash].css')
+        new ExtractTextPlugin('[name].[hash].css'),
+
+        new webpack.optimize.ModuleConcatenationPlugin()
     ]
 });