+++ product/warehouse/src/webapp/src/app/app.module.ts
@@ -10,10 +10,11 @@ import {HomeComponent} from './pages/home/home.component';
 import {HttpModule} from '@angular/http';
 import {InvocationStatusPipe} from './pipes/invocation-status.pipe';
 import {LOCALE_ID, NgModule} from '@angular/core';
-import {LoggedOutComponent} from './pages/logged-out/logged-out.component';
+import {LoggedOutComponent} from './pages/auth/logged-out.component';
 import {MethodDetailComponent} from './pages/methods/method-detail.component';
 import {MethodsComponent} from './pages/methods/methods.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {NotLoggedInComponent} from './pages/auth/not-logged-in.component'
 import {ReportGeneratorComponent} from './pages/report-generator/report-generator.component';
 import {SettingsComponent} from './components/settings-editor.component';
 import {SsoComponent} from './components/sso.component';
@@ -37,6 +38,7 @@ import {IsLoggedIn} from './guards/is-logged-in';
         LoggedOutComponent,
         MethodDetailComponent,
         MethodsComponent,
+        NotLoggedInComponent,
         ReportGeneratorComponent,
         SettingsComponent,
         SsoComponent,