+++ product/warehouse/src/webapp/src/app/guards/is-logged-in.ts
@@ -1,14 +1,19 @@
 import {Injectable} from '@angular/core';
-import {CanActivate} from '@angular/router';
+import {CanActivate, Router} from '@angular/router';
 import {StateService} from '../services/state.service';
 
 @Injectable()
 export class IsLoggedIn implements CanActivate {
 
-    constructor(private stateService: StateService) {
+    constructor(private stateService: StateService, private router: Router) {
     }
 
     canActivate() {
-        return this.stateService.isDemoMode() || this.stateService.isLoggedIn();
+        let result = this.stateService.isDemoMode() || this.stateService.isLoggedIn();
+        if (!result) {
+            // noinspection JSIgnoredPromiseFromCall
+            this.router.navigate(['/not-logged-in']);
+        }
+        return result;
     }
 }