+++ product/warehouse/src/webapp/src/app/pages/auth/not-logged-in.component.ts
@@ -1,0 +1,8 @@
+import {Component} from '@angular/core';
+
+@Component({
+    selector: 'ck-not-logged-in',
+    template: require('./not-logged-in.component.html')
+})
+export class NotLoggedInComponent {
+}