+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 # Codekvast Release Notes
 
 ## 0.21.3
+1. Upgraded to Angular 4.2.4 and Webpack 3.0.0
 
 ## 0.21.2
 1. Added support for Spring Boot executable jars