+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 1. Added column users.lastActivityAt
 1. Now redirects to /logged-out after webapp session has expired
 1. Now redirects to /not-logged-in when an unauthenticated user clicks a link that requires login
+1. Bug fix: now deprovision works.
 
 ## 0.21.2
 1. Added support for Spring Boot executable jars